Create a service/telemetry package, move service::telemetry config
